BUG#33063035 ADD CAPABILITY LIST TO THE DUMP METADATA
Added capability list to the dump metadata.

Shell 8.0.27 has introduced partition awareness to the dump and load
utilities, which required an increase of the major version number of
the dump format. This means that dumps created by Shell 8.0.27 and
newer cannot be loaded by Shell 8.0.26 and older, even if the dump does
not contain partitioned tables.

The capability list will alleviate this problem in the future: if a new
feature is added, and it would have similar impact on the dumper version
as partition awareness, the major version number will remain unchanged.
Instead, a new capability will be added to the list of capabilities and
written do the metadata file, but only if the dump actually uses this
new feature. If an older version of Shell is used to load a dump which
does not use this new capability, the load operation will succeed. On
the other hand, if a dump uses this new capability, and an older version
of Shell does not support it, an error will be reported, and a version
which is capable of loading the dump will be suggested.

Partition awareness has been added as a capability.

Change-Id: Id85b51f74100d69a3e1594e811453bb93479b7d0