Bug#33166390 CLUSTERSET: when doing rejoincluster, a member is kicked out of GR
After a group rebooted, the clusterset_replication channel may not exist
in all members. So when such a cluster is rejoined, the async channel
may be started with auto-failover while some group members do not have
that channel, which causes a GR error.

rejoinCluster() was updated so that the replication settings are updated
in all SECONDARY members first and at the PRIMARY (where the channel is
started) for last.

Change-Id: Id02a9d4093491586568c298056c698de7e90753a