Bug#33166307 ClusterSet: cannot dissolve cluster that was removed from clusterset
Dissolving a cluster that was removed from a clusterset was failing at a
precondition check if its local metadata copy was not synced up during
removal (e.g. because it was force removed while replication was down).

dissolve() was changed to ignore a precondition check errors for
clusters that still belong to a clusterset if the cluster object was
detected to be out-of-sync with its clusterset.

Also: workaround for test failures from GCS port collisions

Change-Id: I5cfd35606547e11f4ed9850c337a8c818aee777a