BUG#33173739 Failed to acquire global read lock
When a global read lock was acquired, dumper checked if user executing
the dump has privileges required to execute the FLUSH TABLES WITH READ
LOCK statement. Depending on the result, FTWRL was executed or a fall
back to LOCK TABLES was used. If FTWRL failed for whatever reason, the
error was reported and dump was aborted. This approach caused problems
on some forks of the MySQL server, as they require additional privileges
to execute FTWRL.

The code has been reworked, if FTWRL fails due to access denied errors,
the fall back to LOCK TABLES is used. In order to keep dry runs
consistent with the normal runs, FTWRL is executed as well, however
session variable lock_wait_timeout is set to a small value to avoid
waiting for long periods of time, while lock, if successfully acquired,
is released immediately.

Change-Id: I81f3b7c45de0423a114ab94786c67d0ff04ae0d8