BUG#28365584 HINT ME WHAT THE "TARGET INSTANCE" IS / GLOBAL SESSION LOSS MESSAGE
This bug is about an error message that is too generic:

Unable to detect target instance state. Please check account
privileges.

When the message is generated for an operation an operation where a
single instance is involved it is clear, however when done in an
operation that involves several instances, the message is ambiguos as
it lacks contexts of the affected instance.

This patch fixes the problem by adding the canonical address of the
affected instance to the error.

Change-Id: I45219913a2b573774d76ce02bde4b84bbbc3be55