BUG#33181308 Shell Dump: incomplete manifest
When Shell was creating a dump with the ociParManifest option set to
true, Shell could appear to have frozen after the summary was shown. If
dump contained a large number of files, it was still creating PARs and
writing the manifest file, but no information was presented to the user.

The code has been updated to wait for the finalization of the manifest
file after the dump is complete but before the summary is printed. While
Shell waits for the manifest to be written, an animated widget is
presented (if showProgress is true) or a message about the beginning and
the end of this operation is displayed.

Change-Id: I4aca8405c3086368ec4162aec43d8c234ee33ea4