BUG#33237648 nonJSON output from addInstance using clone mode
When Shell was started with output in JSON mode, calling addInstance()
method with 'recoveryMethod' set to 'clone' could produce non-JSON
output when Shell detected a server restart and waited for it to come
back.

The widget which is used to display the progress information did not
handle the JSON mode correctly, this is now fixed.

Change-Id: I8563890c5848420a20edb612a48348c6cdaafa5e