BUG#33235685 Shell ignores script arguments when using --interactive option
When a file is processed using the --file(-f) option, any command line
argument after the file name will be passed to the script as system
arguments (argv). This works as expected when the script is being executed
in batch mode, however when the --interactive option is used, the arguments
are not present.

This problem was caused because passing the system arguments was tightly
coupled to the batch execution mode and unavailable for interactive mode.

This patch changes the code to allow setting the system arguments before
the execution is done, independently if it is batch or interactive.

Change-Id: I447df3012b04800e83ecb73f7570c0af304874e7