BUG#32109967 FORCED TO SPECIFY --HOST WHEN USING --SOCKET WITH UTIL.DUMPINSTANCE
When running a dump while user was connected to the server via socket,
operation failed, complaining about missing 'host' connection option.

When dumping the DDL files, Shell includes a comment describing the host
where the MySQL server is running. Code responsible for this mistakenly
assumed that the host information is always available in the connection
options data. This is now fixed, if this information is unavailable
(i.e. when user is connected via socket), host is set to localhost.

Change-Id: I66b3f2a9d304e39f3447bc2f73d7ae6c2bf0f0d5