BUG#33146841 CLUSTERSET: CLUSTERSET OBJECT DOES NOT FOLLOW PRIMARY
Calling <ClusterSet>.setRoutingOption() using a ClusterSet object which
was created while having an established session to a non-primary
instance of the ClusterSet would result in an error indicating the
function was not available through a session to a read only instance.

This was commonly seen when changing the primary instance of a Cluster
that belonged to a ClusterSet.

Even though it seemed that the problem was that the ClusterSet object
was not following the primary changes, the problem was at the
preconditions check.

This patch fixes that by ensuring the preconditions check do not block
the execution of the function when the active session is established to
a R/O (or other state) member, considering that the code acquires the
current primary instance after the preconditions check.

Added new regression tests.

Change-Id: Id2b0c93c1a1a5541a7dffbc43fb7df464ad505f1