BUG#32261381 MYSQL SHELL 8.0.22 LOADDUMP FAILS FOR VIEWS USING FUNCTIONS
If a dump created by the util.dumpTables() function contains i.e. a view
which is using a function, loading it fails with a message like:
  FUNCTION functionName does not exist

This is expected behaviour, since util.dumpTables() function does not
dump routines, any routine which is referenced by the views or triggers
to be dumped is expected to already exist when the dump is loaded.
This limitation was not documented, fixed the docs.

Change-Id: I1c33d87b75d16205a057504196ee2f1e9c8b4b2d