BUG#33223867 dba.rebootClusterFromCompleteOutage() is failing for replica cluster in CS
Using the 'rejoinInstances' option in
dba.rebootClusterFromCompleteOutage() to restore a Replica Cluster, the
command would fail with an error.

The command would not wait for the seed instance to become ONLINE in the
group before attempting to rejoin the remaining instances which is
especially relevant for Replica Clusters on which the managed
replication channel may be started with auto-failover which the members
haven't obtained those channel configurations yet.

This patch fixes that by adding a wait to ensure the seed instance is
ONLINE before rejoining other instances. The timeout for the wait is 5
minutes.

Change-Id: Ic02502010ded8185450e03e8a1cd046fbc7d0c08