BUG#33132299 CLUSTER/CLUSTERSET: WHEN GTID SET IS EMPTY, DEFAULT SHOULD BE CLONE
When Shell detects that an instance being added is new (GTID-SET empty)
it prompts whether Clone based recovery should be used or not, being
that option the default answer. However, if there are missing
transactions that were purged from all Cluster members, the default
answer is to Abort.

In such scenario, the default answer should be Clone as well since the
instance's GTID-SET is empty so is fairly safe to assume the instance is
new and Clone can be used.

For scenarios on which the instance's GTID-set is not empty and
diverging, the default answer remains unchanged being that to Abort.

Updated existing tests to match the change.

Change-Id: I60c5de68576bc9ebebeb91fc38489ab8dffc362f