BUG#33212873 waitDumpTimeout performs concurrent load with negative value
The docs for the util.loadDump() state that the negative value of the
`waitDumpTimeout` option disables the waiting, while it caused a
concurrent load.

The value given by the user was not validated correctly, this is now
fixed.

The `waitDumpTimeout` option was mistakenly documented as an integer,
while it is a float, this has been corrected.

Change-Id: Ibabb78e106ba69ca37e744105e3799b1f295a0eb