BUG#33267266 clusterset: router does not know about global routing option
Router did not consider the global routing option defaults because it
writes to the routers.options table at bootstrap time. Since the
options stored in there have precedence over the global ClusterSet
routing options, those were being ignored.

Router has been changed to not write anything into routers.options when
the bootstrap options regarding routing options are not used, resulting
in the usage of the global options stored in the clustersets table.

For this change to become complete, this patch ensures that the default
global routing options are stored in the clustersets table at ClusterSet
creation time.

On top of that, the view v2_router_options was replaced with a new view
named v2_cs_router_options where clusterset global options defaults are
returned for options that are not overriden by Router. Also reordered
the clusterset_id column to the end of the v2_routers view to keep
backwards compatibility.

Change-Id: I74c7c4d59367030e27a49f66018228e9fc1e9048