BUG#33271037 Object Storage operations sometimes fail with error 401
Sometimes, while dumping or loading to Object Storage, the operation
could fail with a HTTP 401 error:

  The Authorization header has a date that is either too early or too
  late, check that your local clock is correct

When executing HTTP requests to the Object Storage, Shell will
automatically retry the request in case of certain errors. Since the
retry mechanism was reusing the headers with each retry, if problems
persisted for prolonged periods of time, the authorization header would
become invalid. In order to overcome this issue, headers are now
refreshed periodically. Additionally, Shell will now automatically
resend the request in case of HTTP 401 error described above.

Change-Id: I315b2a4a8dc54c0fb0bc500bb7add9d9011f1635