Bug #33267179	clusterset: dissolve() of invalidated/removed cluster still does not work
Fixes for more scenarios for invalidated and removed clusters.

An invalidated PRIMARY cluster that was removed couldn't be dissolved
because it wasn't trying to connect to a different cluster to check if
there was a newer metadata, so from its own point of view, it was still
a valid PRIMARY cluster of the clusterset.

getCluster(), getClusterSet() and rebootCluster were changed to all use
the same checks for clusters that belong to a clusterset and are:
- invalidated (both replica and primary cluster cases)
- removed while OFFLINE

Change-Id: I373d3f7648338a002377424303e8a6cb9aae41bf