BUG#33298735 clusterset: inconsistent case in/sensitivity in clustername
The Cluster name is case sensitive, however, several AdminAPI commands
that require a cluster name as argument treat it as case insensitive.
It's possible to create multiple Replica Clusters in a ClusterSet with
the same name but using different combinations of upper/lower case
characters that ultimately results in unexpected and wrong behavior in
commands such as:

  - setPrimaryCluster()
  - forcePrimaryCluster()
  - setRoutingOption()

This is conceptually wrong since the Cluster name must be unique within
a Metadata schema thus it should be case insensitive.

This patch fixes that by adding the cluster_name as a UNIQUE key in the
Metadata schema and by ensuring the cluster_name comparisons resulting
from Metadata queries are case insensitive, allowing the commands
verification steps to fail before reaching any Metadata update steps.

Change-Id: Iba35d786cc3b15f0ca52eeeeb26968a448c0ba8d