BUG#33267121 clusterset: confusing messaging about async repl during rejoininstance
When adding or rejoining an instance to a Replica Cluster, Shell prints
a message indicating that the replication source of X is being changed
to Y. This message refers to the ClusterSet managed replication channel,
however, that information is not given to the user and can be confusing.

This patch improves that by providing that missing information in the
output.

Change-Id: Ic48e0982742052a4ec64214900a115a779caafa1