BUG#33294003 clusterset: rejoinInstance on REPL wait for sync with PRI, need more verbose out
Rejoining an instance to a Cluster that belongs to a ClusterSet using
.rejoinInstance() does not provide any information about the the
transaction sync that happens under the hood. Depending on the
replication lag, ammount of data to recover, etc. this may take quite
some time misleading the user to think that the command just blocked.

This patch fixes that by providing a more verbose output in the
rejoinInstance() command to indicate when a transaction sync needs to be
done. A progress bar is used when the STDOUT on which Shell is running
is a terminal, otherwise plain textual information is printed.

Change-Id: Iad695cf9cb37ff92df54d05d5a44c968b79fca7c