Post-fix for BUG#33294003:
If the transaction sync used in addInstance()/rejoinInstance()/etc. is
not successfull (due to an exception or because it couldn't sync the
transactions) the output of Shell isn't cleared out becoming unreadable.

The root cause of the issue is that the progress bar isn't terminated in
such scenarios. This patch fixes that by ensuring the progress bar is
terminated in all scenarios.

Change-Id: I984265e7997eb4ae13649d98a2699af73e161570