WL#14244 Extended filtering for dump and load utilities
This patch adds the following new filtering options to the dump and load
utilities:

util.dumpInstance()
===================

includeSchemas
--------------

List of strings (default: empty) - list of schemas to be included in the
dump.

util.dumpInstance() and util.dumpSchemas()
==========================================

includeTables
-------------

List of strings (default: empty) - list of tables or views to be
included in the dump in the format of schema.table.

util.dumpInstance(), util.dumpSchemas() and util.loadDump()
===========================================================

includeRoutines
---------------

List of strings (default: empty) - list of routines to be included in
the dump (loaded from the dump) in the format of schema.routine.

excludeRoutines
---------------

List of strings (default: empty) - list of routines to be excluded from
the dump (not to be loaded from the dump) in the format of
schema.routine.

includeEvents
-------------

List of strings (default: empty) - list of events to be included in the
dump (loaded from the dump) in the format of schema.event.

excludeEvents
-------------

List of strings (default: empty) - list of events to be excluded from the
dump (not to be loaded from the dump) in the format of schema.event.

util.dumpInstance(), util.dumpSchemas(), util.dumpTables() and util.loadDump()
==============================================================================

includeTriggers
---------------

List of strings (default: empty) - list of triggers to be included in the
dump (loaded from the dump) in the format of schema.table (all triggers
from the specified table) or schema.table.trigger (the individual
trigger).

excludeTriggers
---------------

List of strings (default: empty) - list of triggers to be excluded from
the dump (not to be loaded from the dump) in the format of schema.table
(all triggers from the specified table) or schema.table.trigger (the
individual trigger).

Change-Id: Iee76a54e537fc80e5c4dfc4fdc17eda5412f4dff