BUG#33337598 Issuing the \source causes the shell to enter in multiline mode
Script files processed with thw \source shell command may cause the
shell to enter in multiline mode without having any context of what is
expected for the input to complete.

When a file is processed with the \source command it should be expected
that the entire file is successfully executed, letting the shell ready
for the next command, if that is not the case, the shell should raise
an error to indicate what failed in the file processing.

This patch fixes this scenario.

Low Level Changes:
- The Shell_core now contains an interative attribute to decide whether
  the processing should be done in interactive mode or not. This saves
  from having to access the Shell_options in lower layers or to have a
  parameter being passed on the functions.
- Processing a stream now detects if the final state is a statement that
  still needs to continue, if that is the case, it will force executing
  the cached code by disabling the multiline handling, this will cause
  the shell to raise the proper syntax error.
- The Json_shell interface was updated to do similar processing of the
  incoming code, this is, it will force execution disabling multiline
  at the end to have the shell raise the proper errors as well

Change-Id: I1ac0de573df0509efb0433de0fc5d3bfefc933bf