WL#14698 Connection timeout shell options
New connectTimeout Option
=========================

With the introduction of the new `connectTimeout` Shell Option, it will
be used as the default connection timeout for any session out of the
Admin API scope unless a value is explicitly defined when the session is
created.

New dba.connectTimeout Option
=============================

Same as `connectTimeout`, but only affects sessions created for Admin
API operations.

Change-Id: I4c18459d467da5b24d34cf2b0f3a7421df314430