BUG#33301229 cs fails to rejoin the old primary cluster after rebootClusterFromCompleteOutage
Using the options 'removeInstances' and/or 'rejoinInstances' when
rebooting a Cluster that was invalidated in a ClusterSet results
in the introduction of errant transactions in the Cluster.
As a result, the Cluster won't be able to rejoin the ClusterSet
later.

The root cause of this issue is that at the time the cluster is being
rebooted from complete outage, all Metadata changes go through the
instance on which the Cluster is re-bootstrapped since the Cluster is
invalidated.

This patch fixes that by forbidding the usage of those options when the
Cluster is invalidated.

A similar problem also happens when using those options when rebooting
a Replica Cluster: Metadata changes go through the instance on which
the Cluster is re-bootstrapped, however, since the Cluster belongs to
a ClusterSet those changes must happen at the Primary Cluster instead.

This patch fixes that by ensuring the right primary is used in such
scenarios.

Added regressions tests.

Change-Id: Ib8533079717596a15f22b1332337b23041bcd252