BUG#33451028 Unable to register plugin named as pre-exiting plugin nested object
The problem is caused by the basic lookup logic on the Help System at 2 places:

- The help registration logic
- The help Shell Command  (\?)

The registration was improved so when registering extension objects, a lookup
based on the fully qualified id is done. In the past, the only matching done was
based on the new entry name which caused that if another entry named the same was
found it was considered that the Help was already in the system. With this change
the help registration will look for registered help items based on the fully
qualified id, avoiding false positives.

The Shell Command lookup was also updated to give priority to help items having
a fully qualified id equal to the string being used in the search, if no item is
found as an exact match, then the existing selection criteria continues to work.

Change-Id: Iffca1051f84254cf0970f734fbae93632e2eae08