Bug#33462107 plugin_function: unable to attach function to existing object
The issue is caused becayse the python decorators are NOT considering the
name change (for naming convention) when looking for the object hierarchy
where the function is to be added. For that reason the parent object is
"not found" and the logic attempts to register it again, but since
registration logic uses the camelCase naming then it finds the object complains
that the object is already defined.

This patch fixes the object lookup in the registration logic to convert
the name to snake_case using the same algorithm used in the shell so the
object is properly found.

Change-Id: Icb219867e19e4d579492e51f2aa31e5386a3828f