Bug#33468265 shell plugins: unable to get help with object.help()
Two different problems are being fixed with this patch:

The help() function on the objects uses a specific function in the help
system to identify the topic that corresponds to the caller object. This
function was missing logic to do exact match comparison in case multiple
topics are found. This reduces the cases with multiple matches to 0 for
extension objects.

The help system keeps 2 registries of help data: a global context and a
local context. The global context is meant to contain help data for
elements that are available globally (even in threads) while the local
context is for help data to be available in a specific thread. This is
to enable users in different treads have their own set of plugins
(not shared). The problem was that the help for ALL the plugins was loaded
in local context, making it unavailable for other threads even if they
were global plugins. This patch fixes the problem by loading the plugin help
data in global/local context depending on when/where the plugins were loaded.

Change-Id: I5e33860057cc438be41344057dd4bcbae34cad02