Bug#33499602 Windows:wrong handling of prompts when shell is called from external application
When a prompt expecting an answer like Y/N is issued with a shell
started with an external application, the shell is cycling repeating the
question again and again even if the answer provided by the caller
application is valid.

The problem is caused because the routine reading the answer is trimming
the \n character only, in windows it should be trimming \r as well
(\r\n).

This patch fixes the problem by adding the missing trim operation on
windows.

Change-Id: I2d884fe14502077b5a3aedc6a9b4201fd326df46