BUG#33454039 Shell does not compile with Python 3.10 - postfix
When releasing a float object, Python 3.10 checks the interpreter state.
The proxy object which is used to enumerate members by the completion
provider did not ensure that the interpreter state is properly
initialized when the underlying object was released, which caused a
segmentation fault when any of the members was a float. This issue has
been fixed.

Change-Id: I2c41ede743096a0556c0ed3c883328a0c05b7d6e