BUG#33502098 dumpInstance Errors out if the server has just users and there is no other data
When dumping an instance, if server has no schemas, or schema filters
(excludeSchemas, includeSchemas) result in an empty set of schemas,
Shell throws an error and does not allow to dump just the users.

With this fix, if util.dumpInstance() is used and there are no schemas
to be dumped (either they are not present, or schema filters result in
an empty set), but the `users` option is set to `true`, dump will
succeed. If in the same scenario `users` option is set to `false`, Shell
will throw an exception just as previously.

Change-Id: I5323c3ee40a8225b8549a3de44cf19ec8947207e