BUG#33406711 Dump&Load should filter user grants
When using util.dumpInstance() or util.loadDump() to dump or load user
privileges, the GRANT/REVOKE statements which are applied on the
schemas/tables/routines which are excluded from the dump/load operations
are also excluded.

In case of the dump operation, these statements are written, but are
commented out, in case of the load operation, these statements are not
executed.

Statements which apply to the system schemas: information_schema, mysql,
ndbinfo, performance_schema, sys are not affected.

Change-Id: Icfae412c94db1483b68770fe21786b4e00a8f4d1