BUG#33499183: replicaSet.removeInstance() core dumps after replicaSet upgrade
Starting with 8.0.27, the info of the replication accounts used in a
ReplicaSet is stored in the Metadata schema. However, in upgrade
scenarios coming from deployments that used older Shell versions that
data won't be present in the Metadata.

Removing an unreachable instance from a ReplicaSet resulted in a crash
in such scenarios since the logic was attempting to resolve the
replication account info by querying the instance's server_id.

This patch fixes that by ensuring the command does not attempt to query
unreachable instances that do not have the replication account info
stored in the Metadata schema in order to drop the account from the
ReplicaSet, ignoring that step and simply logging it.

Added regression test.

Change-Id: I7a68de35a37032c81e0b2d3072eea5b78d704a6d