BUG#33488614, BUG#33483747, BUG#33470987, BUG#32820283, BUG#32157287
BUG#33488614: typo in cluster.status() when unmanaged replication
channels are found

Fixed typo in cluster.status() that was displaying 'repliication'
instead of 'replication'.

BUG#33483747: Executing rebootClusterFromCompleteOutage() throws
incorrect error message

Attempting to use the options 'removeInstances' and/or 'rejoinInstances'
in dba.rebootClusterFromCompleteOutage() when restoring a Cluster that
is INVALIDATED is forbidden and results in an error. However, that error
message mentions 'removeInstances' twice instead of mentioning the 2
options. This patch fixes that.

BUG#33470987: ClusterSet: improve messaging around FenceWrites..

Attempting to fence from writes a REPLICA cluster results in an
exception, however, no ERROR message is printed which can be confusing
to the end user of the interactive shell.

This patch fixes that by introducing an explanatory error message. Also,
it extendes the function helper text to mention that the function is not
permitted on REPLICA Clusters.

BUG#32820283: GENERAL HELP FOR ADMIN API IS OUTDATED IN SHELL

The general Shell help section was outdated in regards to the AdminAPI.
There were sub-sections not mentioning ReplicaSet or ClusterSet, among
other missing details.

This patch adds and fixes all missing details in the AdminAPI general
help.

BUG#32157287: BUILT-IN HELP FOR DBA.CONFIGURELOCALINSTANCE() IS
MISLEADING REGARDING 1ST ARG

The built-in help for dba.configureLocalInstance() indicates that the
1st parameter (instance) is mandatory. However, that's not true since
it's an optional parameter.

This patch fixes the command help in regards to the command syntax.

Change-Id: Ifc7a5e64643a56c834f1ad9ea9be8a6cc3d94fd4