BUG#33532056: Upgrading shell to 8.0.27 breaks all operations of Cluster running 8.0.25
Clusters running MySQL 8.0.25 are unmanageable with MySQL Shell 8.0.27.
Most operations fail with the preconditions check indicating the Cluster
has lost the quorum when it's perfectly healthy.

This is a frequent issue in upgrade scenarios on which MySQL Shell is
first upgraded to 8.0.27 whilst the Servers running 8.0.25 are only
upgraded afterward.

The root cause of the wrong validation can be traced down to the query
executed to obtain the Cluster's info from the Metadata schema that is
considering that 'group_replication_view_change_uuid' is available since
8.0.25 when it's not. The sysvar was introduced in 8.0.26.

This patch fixes the query accordingly.

Change-Id: I3a608cb02072e9463daa2b0308c45347b8b39fa4