BUG#33402791 util.dump* settings that void one another: includeTables(A) + excludeTables(A)
When using conflicting includeX and excludeX options, both dump and load
utilities will now print an error message, allowing users to pinpoint
the mistake, while the dump/load operation will fail.

The cross checks for the includeTables, includeEvents, includeRoutines,
includeTriggers, excludeTables, excludeEvents, excludeRoutines,
excludeTriggers options have been added as well, if any of the values
refers to a schema which is not included in the dump/load, a similar
error is reported. In case of the includeTriggers and excludeTriggers
options, tables are checked as well. This changes the previous
behaviour, where objects from schemas which were not included in the
dump were not reported as errors, but the intention of this change is to
give a better feedback to the user.

Change-Id: I729ff5afa7c0106df33837b00943aa95ec86e16f