BUG#33492812: Member auto-rejoin with clone fails due to plugin being uninstalled everywhere
The AdminAPI uninstalls the CLONE plugin from all Cluster members when
incremental recovery is used to add a new instance, even though CLONE is
available. This causes any subsequent auto-rejoin that requires CLONE to
fail when it could have succeeded.

This patch ensures the CLONE plugin is not uninstalled whenever
incremental recovery is used. The plugin is installed during
dba.createCluster() and cluster.addInstance() if the instance supports
it and 'disableClone' is not enabled in the Cluster, thereby it
shouldn't be uninstalled to keep it all in sync.

Added regression test.

Change-Id: I615fbe122dae31a968c9a1b7e1d940edb2ffca45