Bug#33389693,Bug#33573834 Can't reboot after total outage with group_seeds set
In some environments, e.g. Kubernetes, the DNS entry of containers
can temporarily disappear when a host is down. This caused clusters
to become unable to bootstrap, because group_replication_group_seeds
can end up having unresolvable hostnames, which causes GR to fail
starting.

That sysvar is now cleared during rebootCluster.

Additionally, the group_seeds handling was fixed to update all cluster
members when instances are rejoined, added and removed. The list of
group_seeds entries was also changed to be taken from the metadata
instead of queried individually.

The groupSeeds option was deprecated and is now ignored in
dba.createCluster() and cluster.addInstance() because it has no known
use cases and was sometimes incorrectly handled.

Bug#33573834 createReplicaCluster() help includes groupSeeds which is not a valid option
Help text for createReplicaCluster() was updated to remove any mentions
of groupSeeds, which was never supported.

Change-Id: I9c1212378fe223068305c7106133493f6377986a