BUG#33551128: From fence Writes to fence All Traffic
Fencing from All Traffic a Cluster that was previously Fenced to Write
traffic was forbidden. However, such operation must be possible
otherwise to perform it one would have to unfence the Cluster first
exposing it to possible unwanted writes.

This patch changes the preconditions of cluster.fenceAllTraffic() to
allow it being executed on Clusters fenced to write traffic.

Added regression tests.

Change-Id: If33b8706bcb174989f668fe962f6b7e71b1e76aa