Fixing help topic registration which causes help tests to fail.
The fix for BUG#33451028 changed the way the help is looked up for
topics using exact topic id matching.

Topic id is created using the fully qualified id of each topic.

After that patch the help for some objects started missing specific
bits, i.e. the help function on the list of functions.

This is caused by a bug on the registration logic which was not
registering the help member with it's fully qualified id.

This patch fixes te registration logic so help topics that get loaded
before their parent topics stay unregistered including the dynamically
agregated help function. Once the link to the root of the help is found
a recursive registration of all the topics with fully qualified id is
perdormed.

Change-Id: Ib5ef0a967e9a14182ab6df483c36af9aee22b7fb