BUG#33445222: ClusterSet.setPrimaryCluster is not blocked with primary OK_FENCED_WRITES status
Attempting to perform a switchover of the Primary Cluster with
.setPrimaryCluster() failed with several errors related to
super_read_only when the target Cluster was Fenced to Write traffic.

This should be avoided since it shouldn't be possible to elect a new
Primary Cluster when the current Primary is fenced to write traffic.

This patch fixes that by ensuring that attempting such operation is
blocked.

Additionally, this patch adds missing preconditions checks to several
ClusterSet operations.

Change-Id: I2e2bd5f6dff9e2860b1affeeccead8c0c3be0096