Fix unit tests
Concurrent dump and load tests on some platforms were failing, because
the order of the dump files copied to simulate a dump in progress is not
fully specified. If the table metadata file was copied before the
trigger SQL file, loader would assume that this table does not have any
triggers, as this was determined after the metadata file was parsed by
checking for the trigger file.

The problem has been fixed, the loader code has been changed to
determine whether or not table has any triggers by using contents of the
metadata file.

Change-Id: I6bf925f7580e3c80a4be1785d83c130dcccd1136