BUG#33414321 index deferment must be disabled for tables with a secondary engine
If a table has a secondary engine defined, it is not possible to execute
DDL statements on such table. The index deferment feature of the
util.loadDump() utility extracts the index definition from the CREATE
TABLE statement, and re-adds them using ALTER TABLE statements once the
data load is finished. If table had a secondary engine, this would fail
with the:
  DDLs on a table with a secondary engine defined are not allowed.
error.

As a fix, the SECONDARY_ENGINE clause is extracted as well, and is added
once all the remaining extracted statements have been executed.

Change-Id: I985e7f28758d81f1ef0245ba36648cd30df1c13a