BUG#33574005: 1:1 mapping of Cluster:Metadata not enforced
When creating a Cluster/ReplicaSet on an instance that already had
a populated Metadata schema (but did not belong to it), the Metadata
was kept and new entries were added to it.

That's wrong since it must not be possible to have multiple Clusters
registered in the same Metadata schema. The 1:1 mapping of
Cluster:Metadata must be enforced, otherwise Shell can't guarantee the
consistency and expected behavior of the supported topologies and opens
the door to unsupported manual attempts to deploy topologies mimicking
what ClusterSet does.

That was the behavior introduced in 8.0.16, however, a regression was
introduced in 8.0.22 that changed again the behavior.

This patch fixes that to ensure that the Metadata schema is dropped and
re-created if already exists.

Added the missing regression tests.

Change-Id: I25bbe452577c4116f4d147e648698d5694dccf85