BUG#33580945: Can not create a Cluster using MySQL Server Version 8.0.25
Attempting to create a cluster on an instance running MySQL 8.0.25 and
using MySQL Shell results in an error.

The root cause of the issue is the check for super_read_only management
that verifies whether the GR member action
mysql_disable_super_read_only_if_primary is enabled on the target
instance. That check is done through a query to a performance_schema
table that was only introduced in 8.0.26, however, the logic was
implemented to run the query when the target instance is >= 8.0.24.

This patch fixes that by ensuring the query is only performed when the
target instance is running a version >= 8.0.26.

Change-Id: I59e3791faa597d37a6d383295fa2e781df5453f3