BUG#33592520 MySQL Shell fix/support --skip-grant-tables
If instance was running with the --skip-grant-tables option enabled,
dumps were failing with the privilege-related errors.

Dump utilities are now able to detect the
'skip-grants user'@'skip-grants host' account, which is assigned to each
connection when the --skip-grant-tables option is enabled, and assume
that this account has all available privileges, resulting in a
successful dump. Additionally, the util.dumpInstance() implicitly sets
the 'users' option to `false`, as some of the grant-related queries fail
when --skip-grant-tables is active.

Change-Id: Ie161e7de8ea2415e99defd5aae6e0a118b263f4f