BUG#33267834: clusterset: removeCluster is listed as independent entity but is really not
ClusterSet.removeCluster() was designed and implemented to ensure the
Cluster is removed from the ClusterSet becoming an independent entity,
i.e. the Cluster becomes a standalone Cluster.

However, the removed Cluster's metadata still contained all the
ClusterSet information that ultimately block any attempt of bootstraping
a Router in the cluster. On top of that, any attempt to perform Cluster
management operations in the Cluster results in unfriendly/confusing
errors since super_read_only is enabled in the whole cluster.

The command was implemented to perform the following tasks in the
following order:

1) Remove Cluster from the ClusterSet members table
2) Drop Replication account
3) Sync to let the changes replicate
4) Stop replication channel and remove ClusterSet related replication
settings
5) Remove the Cluster's metadata from the ClusterSet

To ensure the Cluster really becomes an independent entity, all
ClusterSet information stored in its Metadata must be removed, however,
that results in errant transactions that disallow re-using the Cluster
instances back in the ClusterSet.

The approach taken to solve this scenario is to also implicitly dissolve
the Cluster when removing it from the ClusterSet. This ensures all
members of the removed Cluster become standalone instances but without
any errant transactions.

Change-Id: Ia4a7353ca7def7c303a9c1996208c4361de14709